Challenge 8 - Deliverable 4
Code has error and could not debug.
TypeError: cannot unpack non-iterable NoneType object

My code works in the module practice, but i could not get it to work for this deliverable. if anyone can take a look, that would be greatly appreciated.
Screen shots for this deliverable were from the Module 8 online lessons.